explore.py: Fix collection day counting
They have a bug in the data they give me where some battles
have everyone as performing zero collection day battles.
It is so dumb...  So I put in a workaround to only count if
your collection day battles are less than 3 and not 0, which
is to say that it is identically 1 or 2.