openscad/screwlib: add simplify optional param
This optional parameter for M_screw() and M_nut() will cause the resulting
object to be a simpler version, for example, a cylinder for the shaft instead
of having screw threading.

This is for faster render times when developing and you want to either visualize
with fasteners and increase the number of faces significantly.