arduino/needle-robot/manual-control: add timing to see where time is spent
Looks like a lot of time is spent in Serial.print().  Much of it is actually
writing to the serial port, and I suspect much of it is conversion of integers
to strings.

It is unclear as of now how much time is spent in writing to the serial port
(of which the docs say there is a buffer and Serial.write() instantly returns
if the buffer doesn't fill up - which I queried and it is 63 bytes large)
and how much time is spent in string conversion.

This warrants further experimentation.

I added three headers with this commit:

- TimeReporter.h: Gives the TimeReporter class that manages multiple timers
  by index and will report to the serial port once one timer reaches a
  certain number of timings
- Array.h: similar to std::array giving static sizing but with convenience
  functions and starting out empty, then being able to append and pop much
  like a stack.
- move.h: copy of GCC's std::move() (and std::remove_reference)